b=24465 show extent tree depth and inode number in fsfilt_ext3_get_mblk
Attachment 32952, additional debugging for quota problems we've seen
on nobackupp10. Note that the following portion of the patch was not
applied:

--- a/lustre/quota/quota_context.c
+++ b/lustre/quota/quota_context.c
@@ -310,7 +310,7 @@ check_cur_qunit(struct obd_device *obd,
         limit_org = limit;
         /* when a releasing quota req is sent, before it returned
            limit is assigned a small value. limit will overflow */
-        if (limit + record < 0)
+        if (record < 0)
                 usage -= record;
         else
                 limit += record;

As per b=23858, comment 59 - Johann says

"Seems to make sense at first sight, but would require more testing
imo."

Since this changes functionality beyond adding asserts and printing
debug information, the patch was not included.