b=23858 do not compare unsigned < 0
This is a patch authored by
	Vladimir Saveliev <vladimir.saveliev@oracle.com>.
as a fix to b=23858: Quota code causes run away ll_ost_io thread.

The patch as in attachment 33069 is supposed to be a single patch
solution to b=23858. It included two previous patches in attachments
31912 and 31916. Since I committed both attachments before, this
patch is editted to exclude the earlier changes.

From: Vladimir Saveliev <vladimir.saveliev@oracle.com>
Date: Thu, 6 Oct 2011 05:59:30 +0400
Subject: [PATCH] b=23858 do not compare unsigned < 0

this is also supposed to catch overflow of lqs_bwrite_pending

i=andrew.perepechko
i=johann

	modified:   lustre/quota/quota_master.c