LU-335 ldiskfs: SuSE 11 SP1 update for Lustre 2.X From 1b99f7fbec0c1f472c3799fa9cf673248177dedc Mon Sep 17 00:00:00 2001 From: James Simmons <uja.ornl@gmail.com> Date: Thu, 16 Feb 2012 09:39:29 -0500
The ext4 code difference between SuSE and RHEL6 is very small.
Because of to this ldiskfs support for SuSE servers is mostly
based the RHEL6 ldisk work. Recently Lustre support migrated
to RHEL6.2 support which required the SuSE port to be updated.
This patch provides this update which works for the SuSE
2.6.32.45 kernel as well as remove some no longer needed
ldsiskfs patches for older SuSE releases.

Signed-off-by: James Simmons <uja.ornl@gmail.com>
Change-Id: Ia6763e8a35698018bc84c95564ec9ea63d5f2c33
Reviewed-on: http://review.whamcloud.com/2138
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Tested-by: Hudson
Tested-by: Maloo <whamcloud.maloo@gmail.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>