ORNL-26 prevent call md_set_lock_data() repeatly From 4e54535c857d73f8606d2b5446bad6a6c7a6c532 Mon Sep 17 00:00:00 2001 From: Fan Yong <yong.fan@whamcloud.com> Date: Fri, 4 Nov 2011 00:40:32 +0800 Git-repo: git://git.whamcloud.com/fs/lustre-release.git Git-commit: 4e54535c857d73f8606d2b5446bad6a6c7a6c532
md_set_lock_data() is called from different functions in llite, and
may be called more than once for the same <inode lock> pair. It is
harmless for the correctness, but will affect the performance a bit,
should be avoided.

Drop dentry flags of "DCACHE_LUSTRE_INVALID" only when we hold the
"MDS_INODELOCK_LOOKUP" lock.

Signed-off-by: Fan Yong <yong.fan@whamcloud.com>
Change-Id: I4dbe206af77ba6a619f3268c238ce98ac7aef4c0
Reviewed-on: http://review.whamcloud.com/1224
Tested-by: Hudson
Tested-by: Maloo <whamcloud.maloo@gmail.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Acked-by: Jeff Mahoney <jeffm@suse.com>