LU-9658 ptlrpc: Add QoS for uid and gid in NRS-TBF
This patch add a new QoS feature in TBF policy which could
limits the rate based on uid or gid. The policy is able to
limit the rate both on MDT and OSS site.

The command for this feature is like:
Start the tbf uid QoS:
    lctl set_param ost.OSS.ost_io.nrs_policies="tbf uid"
Limit the rate of ptlrpc requests of the uid 500
    lctl set_param ost.OSS.ost_io.nrs_tbf_rule=
 "start uid={500} rate=100"

Start the tbf gid QoS:
    lctl set_param ost.OSS.ost_io.nrs_policies="tbf gid"
Limit the rate of ptlrpc requests of the gid 500
    lctl set_param ost.OSS.ost_io.nrs_tbf_rule=
 "start gid={500} rate=100"

or use generic tbf rule to mix them:
    lctl set_param ost.OSS.ost_io.nrs_policies="tbf"
Limit the rate of ptlrpc requests of the uid 500 gid 500
    lctl set_param ost.OSS.ost_io.nrs_tbf_rule=
 "start uid={500}&gid={500} rate=100"

Change-Id: I440ad087dd3dbacd8b5228717b0a1724ef47e3b4
Signed-off-by: Teddy Chan <jjkky@yahoo.com>
Signed-off-by: Li Xi <lixi@ddn.com>
Signed-off-by: Wang Shilong <wshilong@ddn.com>
Signed-off-by: Qian Yingjin <qian@ddn.com>