LU-10948 llite: WIP! Introduce inode open counter
We will use this open counter to find inodes that are frequently
opened and closed and use this as a hint to request open lock
during such opens.

Currently the threshold is hardcoded at 10 and we need an interface
exposed to change this value via sysfs or such.

Is there a good reason to have it this high?

Change-Id: I1aa5455b458840acad651f651c883a7a7a67ab4c
Signed-off-by: Oleg Drokin <oleg.drokin@intel.com>