LU-11100: a debug patch
Attached is a debug patch based on the github repo you pointed me to.

This times the lnet_res_lh_lookup() and prints an error message to console if it took more than 5 seconds to finish

When you reproduce the problem, if we can monitor dmesg and see if any of these messages pop up. They are tagged with "NASA_DEBUG".

I'm still scrubbing through the core, printing out the length of each MD hash list to see if there is anyone that is of excessive length. It's taking a while.