LU-12621 o2iblnd: cache max_qp_wr
When creating the device the maximum number of work requests per qp
which can be allocated is already known. Cache that internally,
and when creating the qp make sure the qp's max_send_wr does not
exceed that max. If it does then cap max_send_wr to max_qp_wr.
Recalculate the connection's queue depth based on the max_qp_wr.

Signed-off-by: Amir Shehata <ashehata@whamcloud.com>

[NOTE] This is a back port from b2_12 to 2.10.8, provided by Amir.
	- Jay Lan <jay.j.lan@nasa.gov>